Move CmdExit-based dirs onto rooms for vehicles
